switched api to es6 and implemented express.Routes for routing
